Add files via upload
This project's primary objective is to evaluate salary satisfaction within a given dataset, approaching it as a classification task. The Logistic Regression model was chosen for its effectiveness in such scenarios. The accuracy score, calculated using this model, yielded a favorable result of 0.7044, surpassing the performance of alternative models like k-Nearest Neighbors, Decision Tree, and Random Forest. More details are available in a notebook which is attached to the file.